OP-72188 PyOpsview breaks with multiple systems
When using pyopsview with 2 Opsview systems, creating a connection
to a second system will break the first connection.

The default headers for all connections are set from a specific
dictionary.  However, Python will create a reference between two
dictionaries by default, rather than copy the data.  This results
in the internally stored data being used by all connections created.

The fix is to force the internal dictionary to be copied, not
referenced.